+++ Arcadia/Views/ButtonsLayout/ArcadiaButtonLayout.swift
@@ -1,39 +1,31 @@
 //
-//  GameGearButtonLayout.swift
+//  ArcadiaButtonLayout.swift
 //  Arcadia
 //
-//  Created by Davide Andreoli on 27/07/24.
+//  Created by Davide Andreoli on 04/08/24.
 //
 
 import SwiftUI
 import ArcadiaCore
 
-struct GameGearButtonLayout: View {
+struct ArcadiaButtonLayout: View {
     
     var body: some View {
         HStack {
-            Spacer()
-            CircleButtonView(arcadiaCoreButton: .joypadStart, size: 35)
-        }
-        .padding()
-
-        HStack {
-            DPadView()
             VStack {
                 Spacer()
-                    .frame(maxHeight: 100)
                 CircleButtonView(arcadiaCoreButton: .arcadiaButton, size: 35)
             }
+            .padding()
             Spacer()
-            TwoButtonsView()
         }
-        .padding()
 
-    }
+            
+        }
 }
 
 #Preview {
-    GameGearButtonLayout()
+    ArcadiaButtonLayout()
         .environment(ArcadiaCoreEmulationState.sharedInstance)
         .environment(InputController.shared)
 }